Implemented the HTTP and HTTPS proxy configuration at the charm config

+++ neutron-openvswitch-cplane/hooks/cplane_package_manager.py
@@ -52,8 +52,16 @@ class CPlanePackageManager:
     def _get_pkg_json(self):
         url = self.package_url
         response = None
+        proxies = {}
+        if config('http-proxy'):
+            proxies['http'] = config('http-proxy')
+        if config('https-proxy'):
+            proxies['https'] = config('https-proxy')
         try:
-            response = urllib.urlopen(url)
+            if not proxies:
+                response = urllib.urlopen(url)
+            else:
+                response = urllib.urlopen(url, proxies=proxies)
         except IOError:
             msg = "Invalid URL: URL metioned for Cplane binaries is not valid"
             status_set('blocked', msg)
@@ -143,7 +151,8 @@ package {}".format(version, package_name)
         filename = urlparse.urlsplit(package_dwnld_link).path
         dwnld_package_name = os.path.join(CHARM_LIB_DIR,
                                           os.path.basename(filename))
-        urllib.urlretrieve(package_dwnld_link, dwnld_package_name)
+        if not os.path.exists(dwnld_package_name):
+            urllib.urlretrieve(package_dwnld_link, dwnld_package_name)
 
         if self.verify_file_checksum(dwnld_package_name, file_checksum):
             juju_log("Package %s downloaded successfully"
